requirements.txt
deepseek
beautifulsoup4
requests
markdown
pydantic
python-dotenv
